♻️ Refactor: 파일 검증 및 프리뷰 생성 util 함수로 분리
